+++ src/core/server/extensions/playerFunctions/select.ts
@@ -124,9 +124,8 @@ async function selectCharacter(player: alt.Player, characterData: Partial<Charac
         // Finish Selection
         alt.emit(ATHENA_EVENTS_PLAYER.SELECTED_CHARACTER, player);
         playerFuncs.set.frozen(player, false);
+        player.visible = true;
     }, 500);
-
-    player.visible = true;
 
     // Delete unused data from the Player.
     delete player.currentCharacters;